List group and unit test wrapper component updates
- Pulled unit test results list item into SB
- All unique aspects of UT results are bundled in the IC which extend base item
- Cleansed UT wrapper of inline components
- Setting up list group and list group item base for middle area
