+++ packages/insomnia-app/app/ui/components/wrapper-unit-test.js
@@ -1,6 +1,5 @@
 // @flow
 import * as React from 'react';
-import styled from 'styled-components';
 import autobind from 'autobind-decorator';
 import classnames from 'classnames';
 import PageLayout from './page-layout';
@@ -13,6 +12,10 @@ import {
   SvgIcon,
   ListGroup,
   ListGroupItem,
+  UnitTestResultItem,
+  StyledFailedBadge,
+  StyledPassedBadge,
+  StyledTimestamp,
 } from 'insomnia-components';
 import ErrorBoundary from './error-boundary';
 import CodeEditor from './codemirror/code-editor';
@@ -42,72 +45,6 @@ type State = {|
   testsRunning: Array<UnitTest> | null,
   resultsError: string | null,
 |};
-
-const StyledTimestamp: React.ComponentType<{}> = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-start;
-  flex-shrink: 0;
-  font-size: var(--font-size-xs);
-  color: var(--hl-xl);
-
-  svg {
-    fill: var(--hl-xl);
-    margin-right: var(--padding-xxs);
-  }
-`;
-
-const StyledBadge: React.ComponentType<{}> = styled.span`
-  padding: var(--padding-xs) var(--padding-sm);
-  border: 1px solid var(--color-success);
-  background-color: var(--color-bg);
-  color: var(--color-success);
-  font-weight: var(--font-weight-bold);
-  border-radius: var(--radius-sm);
-  flex-basis: 2.5em;
-  flex-shrink: 0;
-  text-align: center;
-  text-transform: capitalize;
-`;
-
-const StyledFailedBadge = styled(StyledBadge)`
-  border-color: var(--color-danger);
-  color: var(--color-danger);
-`;
-
-const StyledPassedBadge = styled(StyledBadge)`
-  border-color: var(--color-success);
-  color: var(--color-success);
-`;
-
-const StyledResultListItem: React.ComponentType<{}> = styled.li`
-  > div:first-of-type {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-between;
-
-    > *:not(:first-child) {
-      margin: var(--padding-xs) 0 var(--padding-xs) var(--padding-sm);
-    }
-  }
-
-  code {
-    background-color: var(--hl-xs);
-    padding: var(--padding-sm) var(--padding-md) var(--padding-sm) var(--padding-md);
-    color: var(--color-danger);
-    display: block;
-    margin: var(--padding-sm) 0 0 0;
-  }
-
-  p {
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    flex-grow: 1;
-  }
-`;
 
 @autobind
 class WrapperUnitTest extends React.PureComponent<Props, State> {
@@ -405,23 +342,21 @@ class WrapperUnitTest extends React.PureComponent<Props, State> {
             </div>
             <ListGroup>
               {tests.map((t, i) => (
-                <ListGroupItem key={i}>
-                  <StyledResultListItem>
-                    <div>
-                      {t.err.message ? (
-                        <StyledFailedBadge>Failed</StyledFailedBadge>
-                      ) : (
-                        <StyledPassedBadge>Passed</StyledPassedBadge>
-                      )}
-                      <p>{t.title}</p>
-                      <StyledTimestamp>
-                        <SvgIcon icon="clock" />
-                        <span>{t.duration} ms</span>
-                      </StyledTimestamp>
-                    </div>
-                    {t.err.message && <code>{t.err.message}</code>}
-                  </StyledResultListItem>
-                </ListGroupItem>
+                <UnitTestResultItem key={i}>
+                  <div>
+                    {t.err.message ? (
+                      <StyledFailedBadge>Failed</StyledFailedBadge>
+                    ) : (
+                      <StyledPassedBadge>Passed</StyledPassedBadge>
+                    )}
+                    <p>{t.title}</p>
+                    <StyledTimestamp>
+                      <SvgIcon icon="clock" />
+                      <span>{t.duration} ms</span>
+                    </StyledTimestamp>
+                  </div>
+                  {t.err.message && <code>{t.err.message}</code>}
+                </UnitTestResultItem>
               ))}
             </ListGroup>
           </div>
@@ -437,7 +372,7 @@ class WrapperUnitTest extends React.PureComponent<Props, State> {
     const selectableRequests = this.buildSelectableRequests();
 
     return (
-      <div key={unitTest._id} className="unit-tests__tests__block">
+      <ListGroupItem key={unitTest._id}>
         <div className="unit-tests__tests__block__header">
           <h2 className="pad-left-md">
             <Editable
@@ -493,7 +428,7 @@ class WrapperUnitTest extends React.PureComponent<Props, State> {
           lineWrapping={settings.editorLineWrapping}
           placeholder=""
         />
-      </div>
+      </ListGroupItem>
     );
   }
 
@@ -532,7 +467,7 @@ class WrapperUnitTest extends React.PureComponent<Props, State> {
               {testsRunning ? 'Running... ' : 'Run Tests'}
             </Button>
           </div>
-          {activeUnitTests.map(this.renderUnitTest)}
+          <ListGroup>{activeUnitTests.map(this.renderUnitTest)}</ListGroup>
         </div>
         {this.renderResults()}
       </div>
