+++ packages/insomnia-components/components/list-group/list-group-item.js
@@ -1,0 +1,7 @@
+// @Flow
+import styled from 'styled-components';
+
+export default styled.li`
+  border-bottom: 1px solid var(--hl-xs);
+  padding: var(--padding-sm) var(--padding-sm);
+`;
