+++ packages/insomnia-components/index.js
@@ -7,8 +7,8 @@ import _DropdownDivider from './components/dropdown/dropdown-divider';
 import _DropdownItem from './components/dropdown/dropdown-item';
 import _GravatarImg from './components/gravatar-img';
 import _Header from './components/header';
-import _ListGroup from './components/list-group';
-import _ListGroupItem from './components/list-group-item';
+import _ListGroup from './components/list-group/list-group';
+import _ListGroupItem from './components/list-group/list-group-item';
 import _MultiSwitch from './components/multi-switch';
 import _NoticeTable from './components/notice-table';
 import _SvgIcon from './components/svg-icon';
@@ -16,6 +16,7 @@ import _Sidebar from './components/sidebar/';
 import _RadioButtonGroup from './components/radio-button-group';
 import _Switch from './components/switch';
 import _ToggleSwitch from './components/toggle-switch';
+import * as unitTestResultItem from './components/list-group/unit-test-result-item';
 import * as table from './components/table';
 
 export const Breadcrumb = _Breadcrumb;
@@ -29,6 +30,10 @@ export const GravatarImg = _GravatarImg;
 export const Header = _Header;
 export const ListGroup = _ListGroup;
 export const ListGroupItem = _ListGroupItem;
+export const UnitTestResultItem = unitTestResultItem.UnitTestResultItem;
+export const StyledFailedBadge = unitTestResultItem.StyledFailedBadge;
+export const StyledPassedBadge = unitTestResultItem.StyledPassedBadge;
+export const StyledTimestamp = unitTestResultItem.StyledTimestamp;
 export const MultiSwitch = _MultiSwitch;
 export const NoticeTable = _NoticeTable;
 export const RadioButtonGroup = _RadioButtonGroup;
